requirements.txt
claude-agent-sdk>=0.1.52
feedparser>=6.0.0
sgmllib3k>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
